heczn/requirements.txt
PySide6>=6.4.0
psutil>=5.8.0
pillow>=9.0.0
pytesseract>=0.3.10
googletrans==4.0.0-rc1